refactor and implement writer
