+++ pkg/writer.go
@@ -14,24 +14,38 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package eth_access_lists
+package migration_tools
 
 import (
 	"github.com/jmoiron/sqlx"
-	"github.com/vulcanize/migration-tools/pkg/interfaces"
 )
 
-// Writer struct for writing v3 DB eth.access_list_elements models
+// Writer struct for writing v3 DB public.nodes models
 type Writer struct {
 	db *sqlx.DB
 }
 
-// NewWriter satisfies interfaces.WriterConstructor for eth.access_list_elements
-func NewWriter(db *sqlx.DB) interfaces.Writer {
+// NewWriter satisfies interfaces.WriterConstructor for public.nodes
+func NewWriter(db *sqlx.DB) *Writer {
 	return &Writer{db: db}
 }
 
-// Write satisfies interfaces.Writer for eth.access_list_elements
-func (w *Writer) Write(models [][]interface{}) error {
-
+// Write satisfies interfaces.Writer for public.nodes
+func (w *Writer) Write(pgStr WritePgStr, models interface{}) error {
+	tx, err := w.db.Beginx()
+	if err != nil {
+		return err
+	}
+	defer func() {
+		if p := recover(); p != nil {
+			rollback(tx)
+			panic(p)
+		} else if err != nil {
+			rollback(tx)
+		} else {
+			err = tx.Commit()
+		}
+	}()
+	_, err = tx.NamedQuery(string(pgStr), models)
+	return err
 }
